[WINESYNC] adsldp/tests: Skip all remaining tests when the server is down.
There's a long timeout when the connection fails and there doesn't
appear to be a way to reduce it with this API. When several of these
timeouts occur the total time can exceed the timeout for the test.

This patch skips all remaining tests when one of them detects that
the server is down.

wine commit id 39e70374e67a84ee96a660b637a6c148d0513c91 by Hans Leidekker <hans@codeweavers.com>